dialog transition set; check file redirection

+++ src/js/init.js
@@ -1,11 +1,13 @@
 import fs from 'fs'
 import path from 'path'
 export default () => {
-    try {
-        fs.readFileSync(path.join(process.cwd(), '_check'))
-    } catch (error) {
-        console.log('localStorage clear')
-        localStorage.clear()
-        fs.writeFileSync(path.join(process.cwd(), '_check'), `check at ${new Date()}`)
+    if (process.env.APPENV !== 'DEV') {
+        try {
+            fs.readFileSync(path.join(process.cwd(), '..', '.yexo_check'))
+        } catch (error) {
+            console.log('localStorage clear')
+            localStorage.clear()
+            fs.writeFileSync(path.join(process.cwd(), '..', '.yexo_check'), `check at ${new Date()}`);
+        }
     }
 }
